+++ @pageobject/class/src/__mocks__/_findElement.ts
@@ -1,6 +1,6 @@
 const findElementModule = jest.genMockFromModule<{
   findElement: jest.Mock;
-}>('../findElement');
+}>('../_findElement');
 
 findElementModule.findElement = jest.fn();