+++ @pageobject/puppeteer-adapter/src/__tests__/PuppeteerAdapter.test.ts
@@ -1,68 +1,36 @@
-import {join} from 'path';
-import {PuppeteerAdapter, predicates} from '..';
-import {IndexPage} from '../__fixtures__/IndexPage';
+import {AdapterTestSuite, PageObject} from '@pageobject/class';
+import {ElementHandle} from 'puppeteer';
+import {PuppeteerAdapter} from '..';
 
-const url = `file://${join(__dirname, '../__fixtures__/index.html')}`;
+class TestPage extends PageObject<ElementHandle, PuppeteerAdapter> {
+  public static selectors: string[];
+  public static url: RegExp;
+}
 
-let adapter: PuppeteerAdapter;
+class PuppeteerAdapterTestSuite extends AdapterTestSuite<
+  ElementHandle,
+  PuppeteerAdapter
+> {
+  public async open(url: string): Promise<PuppeteerAdapter> {
+    const adapter = await PuppeteerAdapter.launchChrome();
 
-beforeEach(async () => {
-  adapter = await PuppeteerAdapter.launchChrome();
+    TestPage.selectors = this.initialSelectors;
+    TestPage.url = this.urlPattern;
 
-  await adapter.open(IndexPage, url);
-});
+    await adapter.open(TestPage, url);
 
-afterEach(async () => {
-  await adapter.browser.close();
-});
+    return adapter;
+  }
 
-describe('SeleniumAdapter', () => {
-  describe('public this.findElements(selector, parent?)', () => {
-    it('should return all <p> elements', async () => {
-      const elements = await adapter.findElements('p');
+  public async quit(adapter: PuppeteerAdapter): Promise<void> {
+    await adapter.browser.close();
+  }
+}
 
-      expect(elements).toHaveLength(2);
+describe('PuppeteerAdapter', () => {
+  it('should pass the test suite', async () => {
+    const testSuite = new PuppeteerAdapterTestSuite();
 
-      expect(
-        await predicates.textEquals('Foo')(adapter, elements[0], 0, elements)
-      ).toBe(true);
-
-      expect(
-        await predicates.textEquals('Bar')(adapter, elements[1], 1, elements)
-      ).toBe(true);
-    });
-
-    it('should return only the descendant <p> element of #bar', async () => {
-      const container = (await adapter.findElements('#bar'))[0];
-      const elements = await adapter.findElements('p', container);
-
-      expect(elements).toHaveLength(1);
-
-      expect(
-        await predicates.textEquals('Bar')(adapter, elements[0], 0, elements)
-      ).toBe(true);
-    });
-
-    it('should return no elements', async () => {
-      const elements = await adapter.findElements('.undefined');
-
-      expect(elements).toHaveLength(0);
-    });
-  });
-
-  describe('public this.getCurrentUrl()', () => {
-    it('should return the initial url', async () => {
-      expect(await adapter.getCurrentUrl()).toBe(url);
-    });
-
-    it('should return a manipulated url', async () => {
-      const manipulatedUrl = `${url}?foo=bar`;
-
-      await adapter.page.evaluate((_url: string) => {
-        window.history.pushState({}, 'Test', _url);
-      }, manipulatedUrl);
-
-      expect(await adapter.getCurrentUrl()).toBe(manipulatedUrl);
-    });
+    await testSuite.runTests();
   });
 });